+++ CHANGELOG.md
@@ -2,16 +2,27 @@
 
 *Changelog created using the [Simple Changelog](https://marketplace.visualstudio.com/items?itemName=tobiaswaelde.vscode-simple-changelog) extension for VS Code.*
 
+## [1.0.6] - 2023-03-26
+
+### Added
+
+- Chapter 6: Surface Normals and Multiple Objects
+- Hit point with the sphere
+- Find the normal vector of the hit point
+- Refactor the code to use a `Hittable` interface
+- add rtweekend.hpp to the project
+
 ## [1.0.5] - 2023-03-26
 
-### Added (1.0.5)
+### Added
 
 - Chapter 5: Adding a Sphere
 - added function to detect collision between ray and sphere
 
+
 ## [1.0.4] - 2023-03-26
 
-### Added (1.0.4)
+### Added
 
 - Initial release
 - Chapter 4: Rays, a Simple Camera, and Background